requirements.txt
Flask==3.0.3
pytest==8.2.2
python-dotenv==1.0.1
requests==2.32.3
twilio==9.2.1

